# api/swagger.yaml
basePath: /
definitions:
...
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
...
  model.People:
    properties:
      alternativeNames:
        items:
          type: string
        type: array
      birthday:
        $ref: '#/definitions/model.Date'
      familyName:
        type: string
      favorite:
        type: integer
      givenName:
        type: string
      id:
        type: integer
      image:
        type: string
      more:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
...
  model.Role:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  model.VoiceActor:
    properties:
      anime:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: MyAnimeList database dump and API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: MAL-DB API
paths:
...
  /people/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.People'
              type: object
      summary: Get people details
      tags:
      - people
  /people/{id}/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get people published manga list
      tags:
      - people
  /people/{id}/staff:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get people staff role list
      tags:
      - people
  /people/{id}/va:
    get:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.VoiceActor'
                  type: array
              type: object
      summary: Get people character role list
      tags:
      - people
...
schemes:
- http
- https
swagger: "2.0"